chore: IHttpParserHelper moved and HttpParserHelper namespace updated
